♻️ useLanguage hook
